refactor: 🧩 Change activity history to scoped styling
