+++ src/pages/User/Overview/ActivityHistory.scoped.css
@@ -1,0 +1,89 @@
+@layer layout {
+  .activity-history-container {
+    display: grid;
+    grid-template-columns: auto 1fr;
+    gap: .25rem;
+    padding: .5rem;
+
+    .activity-history-header-list {
+      display: grid;
+      list-style: none;
+      padding: 0;
+      align-items: center;
+
+      .activity-history-header {
+        line-height: 0;
+        font-size: .9em;
+      }
+    }
+
+    .activity-history-list {
+      display: grid;
+      gap: .15rem;
+      padding: 0;
+      grid-template-rows: repeat(7, auto);
+      grid-auto-flow: column;
+      grid-auto-columns: 1fr;
+      list-style: none;
+
+      .activity-item {
+        position: relative;
+        aspect-ratio: 1/1;
+        border-radius: .2rem;
+        background-color: grey;
+
+        &[data-level="1"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), black 30%);
+        }
+        &[data-level="2"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), black 20%);
+        }
+        &[data-level="3"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), black 10%);
+        }
+        &[data-level="4"] {
+          background-color: var(--user-color, orange);
+        }
+        &[data-level="5"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), white 10%);
+        }
+        &[data-level="6"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), white 20%);
+        }
+        &[data-level="7"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), white 20%);
+        }
+        &[data-level="8"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), white 30%);
+        }
+        &[data-level="9"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), white 40%);
+        }
+        &[data-level="10"] {
+          background-color: color-mix(in hsl, var(--user-color, orange), white 50%);
+        }
+      }
+    }
+
+    @media (width < 55em) {
+      padding: 1rem;
+      gap: .25rem;
+    }
+
+    @media (width < 35em) {
+      padding: .5rem;
+      gap: .15rem;
+    }
+
+    @media (width < 22em) {
+      padding: .5rem;
+      gap: .15rem;
+
+      .activity-history-list {
+        gap: 0.05rem;
+        border-radius: .05rem;
+      }
+    }
+  }
+}
+
