+++ src/pages/User/Overview/Overview.jsx
@@ -1,13 +1,12 @@
-import { A } from "@solidjs/router";
-import { createMemo, For, Match, Show, Switch } from "solid-js";
+import {A} from "@solidjs/router";
+import {For, Match, Show, Switch} from "solid-js";
 import api from "../../../utils/api.js";
-import { useAuthentication, useUser } from "../../../context/providers.js";
-import { formatTimeToDate, formatTitleToUrl, mediaUrl, numberCommas } from "../../../utils/formating.js";
-import { ActivityCard } from "../../../components/Activity.jsx";
-import { Tooltip } from "../../../components/Tooltips.jsx";
+import {useAuthentication, useUser} from "../../../context/providers.js";
+import {formatTitleToUrl, mediaUrl, numberCommas} from "../../../utils/formating.js";
+import {ActivityCard} from "../../../components/Activity.jsx";
 import "./Overview.scss";
-import { asserts, timeCollection } from "../../../collections/collections.js";
-import { arrayUtils, numberUtils } from "../../../utils/utils.js";
+import {asserts} from "../../../collections/collections.js";
+import {ActivityHistoryScoped} from "./ActivityHistory.scoped.jsx";
 
 export function Overview() {
   const { user } = useUser();
@@ -28,7 +27,7 @@ export function Overview() {
             <p>{user().donatorBadge} (extra fancy)</p>
           </Match>
         </Switch>
-        <ActivityHistory history={user().stats?.activityHistory || []} />
+        <ActivityHistoryScoped history={user().stats?.activityHistory || []} />
         <Show when={user().favourites.anime.edges.length}>
           <div class="user-favourites">
             <h3>Favourite animes</h3>
@@ -179,90 +178,4 @@ function GenrePreview(props) {
       </div>
     </Show>
   );
-}
-
-const firstDay = 1 // Monday
-const createStartDate = () => {
-  const [date] = new Date().toISOString().split("T");
-  const start = new Date(`${date}T00:00`);
-  // Move start half a year
-  start.setDate(start.getDate() - 182);
-  // Set start to the firstDay of that week
-  start.setDate((start.getDate() - start.getDay()) + firstDay);
-  return start.getTime();
-}
-
-function ActivityHistory(props) {
-  const start = createStartDate() / 1000;
-
-  const days = createMemo(() => {
-    const end = new Date() / 1000;
-    const parsedHistory = [];
-
-    props.history?.forEach((cur, i, arr) => {
-      const delta = cur.date - numberUtils.max(arr[i - 1]?.date, start - timeCollection.dayAsSeconds);
-      // Date is out of range skip iteration
-      if (cur.date < start || cur.date > end + timeCollection.dayAsSeconds) {
-        return;
-      }
-
-      // Add missing empty dates in between history
-      for (let i = Math.floor(delta / timeCollection.dayAsSeconds) - 1; i > 0; i--) {
-        parsedHistory.push({ date: cur.date - timeCollection.dayAsSeconds * i });
-      }
-
-      parsedHistory.push(cur);
-    });
-
-    // Add missing empty dates at the end
-    const lastDate = props.history?.at(-1)?.date;
-    const rightPadding = Math.floor((end - lastDate) / timeCollection.dayAsSeconds);
-    for (let i = 1; i < rightPadding; i++) {
-      parsedHistory.push({ date: lastDate + timeCollection.dayAsSeconds * i });
-    }
-
-    return parsedHistory;
-  });
-
-  const dayNames = [ "Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat" ];
-
-  return (
-    <Show when={props.history.at(-1).date > start}>
-      <div>
-        <h3>Activity</h3>
-        <div class="activity-history-container">
-          <ol class="activity-history-header-list">
-            <For each={dayNames}>{(_, i) => (
-              <li class="activity-history-header">{arrayUtils.at(dayNames, i() + firstDay)}</li>
-            )}</For>
-          </ol>
-          <ol class="activity-history-list">
-            <For each={days()}>{(activity) => (
-              <li class="activity-item" attr:data-level={activity.level}>
-                <ActivityTooltip date={activity.date * 1000} amount={activity.amount} />
-              </li>
-            )}</For>
-          </ol>
-        </div>
-      </div>
-    </Show>
-  );
-
-  function ActivityTooltip(props) {
-    const getTipPosition = date => {
-      if (date < (start + 3600 * 24 * 60) * 1000) {
-        return "right";
-      }
-      if (date > (start + 3600 * 24 * (180 - 60)) * 1000) {
-        return "left";
-      }
-    }
-
-    return (
-      <Tooltip tipPosition={getTipPosition(props.date)}>
-        <p>{formatTimeToDate(props.date)}</p>
-        <p>Amount: {props.amount || 0}</p>
-      </Tooltip>
-    );
-  }
 }
